refactor(date-time): centralize date formatting logic into a utility function for improved localization support
