+++ spring-ai-alibaba-admin/frontend/packages/main/src/legacy/pages/evaluation/experiment/experimentDetail/index.tsx
@@ -4,6 +4,7 @@ import { Button, Card, Progress, Tag, Alert, Spin, Table, message, Typography, T
 import { ArrowLeftOutlined, StopOutlined, ReloadOutlined } from '@ant-design/icons';
 import API from '../../../../services';
 import usePagination from '../../../../hooks/usePagination';
+import { formatDateTime } from '../../../../utils/formatDateTime';
 import './index.css';
 import $i18n from '@/i18n';
 
@@ -19,25 +20,6 @@ const scrollbarHideStyle = `
     scrollbar-width: none;
   }
 `;
-
-// 格式化时间显示
-const formatDateTime = (dateTimeString: string) => {
-    if (!dateTimeString) return '-';
-    try {
-        const date = new Date(dateTimeString);
-        const localeMap: Record<string, string> = { zh: 'zh-CN', en: 'en-US', ja: 'ja-JP' };
-        return date.toLocaleString(localeMap[$i18n.getCurrentLanguage()] || 'en-US', {
-            year: 'numeric',
-            month: '2-digit',
-            day: '2-digit',
-            hour: '2-digit',
-            minute: '2-digit',
-            second: '2-digit'
-        });
-    } catch {
-        return dateTimeString;
-    }
-};
 
 // 评测结果数据接口
 interface EvaluationResult {
