+++ spring-ai-alibaba-admin/frontend/packages/main/src/legacy/pages/evaluation/experiment/index.tsx
@@ -7,30 +7,12 @@ import API from '../../../services';
 import ExperimentCreate from './experimentCreate';
 import usePagination from '../../../hooks/usePagination';
 import { getLegacyPath } from '../../../utils/path';
+import { formatDateTime } from '../../../utils/formatDateTime';
 import './index.css';
 import $i18n from '@/i18n';
 
 const { Option } = Select;
 const { Title } = Typography;
-
-// 格式化时间显示
-const formatDateTime = (dateTimeString: string) => {
-    if (!dateTimeString) return '-';
-    try {
-        const date = new Date(dateTimeString);
-        const localeMap: Record<string, string> = { zh: 'zh-CN', en: 'en-US', ja: 'ja-JP' };
-        return date.toLocaleString(localeMap[$i18n.getCurrentLanguage()] || 'en-US', {
-            year: 'numeric',
-            month: '2-digit',
-            day: '2-digit',
-            hour: '2-digit',
-            minute: '2-digit',
-            second: '2-digit'
-        });
-    } catch {
-        return dateTimeString;
-    }
-};
